feat(frontend): hide brief if it is empty
